+++ src/pages/entry/ObservationEntry/ObservationEntry.tsx
@@ -44,11 +44,12 @@ import {
   FREQUENCY_MHZ,
   FREQUENCY_GHZ,
   FOOTER_SPACER,
-  WRAPPER_HEIGHT
+  WRAPPER_HEIGHT,
+  TYPE_ZOOM
 } from '../../../utils/constants';
 import HelpPanel from '../../../components/info/helpPanel/helpPanel';
 import Proposal from '../../../utils/types/proposal';
-import { frequencyConversion, generateId } from '../../../utils/helpers';
+import { frequencyConversion, generateId, getMinimumChannelWidth } from '../../../utils/helpers';
 import AddButton from '../../../components/button/Add/Add';
 import ImageWeightingField from '../../../components/fields/imageWeighting/imageWeighting';
 import GroupObservationsField from '../../../components/fields/groupObservations/groupObservations';
@@ -63,8 +64,8 @@ import RobustField from '../../../components/fields/robust/Robust';
 import SpectralAveragingField from '../../../components/fields/spectralAveraging/SpectralAveraging';
 import SpectralResolutionField from '../../../components/fields/spectralResolution/SpectralResolution';
 import NumStations from '../../../components/fields/numStations/NumStations';
-import ContinuumBandwidthField from '../../../components/fields/continuumBandwidth/continuumBandwidth';
-import BandwidthField from '../../../components/fields/bandwidth/bandwidth';
+import ContinuumBandwidthField from '../../../components/fields/bandwidthFields/continuumBandwidth/continuumBandwidth';
+import BandwidthField from '../../../components/fields/bandwidthFields/bandwidth/bandwidth';
 
 const TOP_LABEL_WIDTH = 6;
 const BOTTOM_LABEL_WIDTH = 6;
@@ -113,6 +114,8 @@ export default function ObservationEntry() {
   const [numOf13mAntennas, setNumOf13mAntennas] = React.useState(0);
   const [numOfStations, setNumOfStations] = React.useState(0);
   const [validateToggle, setValidateToggle] = React.useState(false);
+  const [scaledBandwidth, setScaledBandwidth] = React.useState<number>(0);
+  const [minimumChannelWidthHz, setMinimumChannelWidthHz] = React.useState<number>(0);
 
   const [groupObservation, setGroupObservation] = React.useState(0);
   const [myObsId, setMyObsId] = React.useState('');
@@ -324,7 +327,6 @@ export default function ObservationEntry() {
     }
   };
 
-  // TODO expand on this to fix 5a band default value issue when switching from Low
   React.useEffect(() => {
     const calculateSubarray = () => {
       if (isEdit()) {
@@ -339,8 +341,12 @@ export default function ObservationEntry() {
       }
     };
 
+    const calculateMinimumChannelWidthHz = () =>
+      setMinimumChannelWidthHz(getMinimumChannelWidth(telescope()));
+
     calculateSubarray();
     setFrequencyUnits();
+    calculateMinimumChannelWidthHz();
   }, [observingBand]);
 
   const isContinuum = () => observationType === TYPE_CONTINUUM;
@@ -705,7 +711,7 @@ export default function ObservationEntry() {
     return fieldWrapper(
       <ContinuumBandwidthField
         labelWidth={BOTTOM_LABEL_WIDTH}
-        onFocus={() => helpComponent(t('continuumBandWidth.help'))}
+        onFocus={() => helpComponent(t(`bandwidth.help.${TYPE_CONTINUUM}`))}
         setValue={setContinuumBandwidth}
         value={continuumBandwidth}
         suffix={continuumBandwidthUnitsField()}
@@ -715,22 +721,33 @@ export default function ObservationEntry() {
         centralFrequency={centralFrequency}
         centralFrequencyUnits={centralFrequencyUnits}
         subarrayConfig={subarrayConfig}
+        setScaledBandwidth={setScaledBandwidth}
+        minimumChannelWidthHz={minimumChannelWidthHz}
       />
     );
   };
 
-  const bandwidthField = () =>
-    fieldWrapper(
-      <BandwidthField
-        onFocus={() => helpComponent(t('bandwidth.help'))}
-        required
-        setValue={setBandwidth}
-        testId="bandwidth"
-        value={bandwidth}
-        telescope={telescope()}
-        widthLabel={BOTTOM_LABEL_WIDTH}
-      />
-    );
+  const bandwidthField = () => (
+    <Grid item>
+      {fieldWrapper(
+        <BandwidthField
+          onFocus={() => helpComponent(t(`bandwidth.help.${TYPE_ZOOM}`))}
+          required
+          setValue={setBandwidth}
+          testId="bandwidth"
+          value={bandwidth}
+          telescope={telescope()}
+          widthLabel={BOTTOM_LABEL_WIDTH}
+          observingBand={observingBand}
+          centralFrequency={centralFrequency}
+          centralFrequencyUnits={centralFrequencyUnits}
+          subarrayConfig={subarrayConfig}
+          setScaledBandwidth={setScaledBandwidth}
+          minimumChannelWidthHz={minimumChannelWidthHz}
+        />
+      )}
+    </Grid>
+  );
 
   const spectralResolutionField = () =>
     fieldWrapper(
@@ -848,11 +865,26 @@ export default function ObservationEntry() {
     );
   };
 
+  // SARAH
   const SubBandsField = () => {
     const errorMessage = () => {
       const min = Number(t('subBands.range.lower'));
       const max = Number(t('subBands.range.upper'));
-      return subBands < min || subBands > max ? t('subBands.range.error') : '';
+      if (subBands < min || subBands > max) {
+        return t('subBands.range.error');
+      }
+      // The sub-band bandwidth defined by the bandwidth of the observation divided by the number of
+      // sub-bands should be greater than the minimum allowed bandwidth
+      if (!isLow() && isContinuum()) {
+        if (
+          scaledBandwidth !== 0 &&
+          subBands &&
+          scaledBandwidth / subBands < minimumChannelWidthHz
+        ) {
+          return t('subBands.range.bandwidthSubBand');
+        }
+      }
+      return '';
     };
 
     const validate = (e: number) => {